requirements.txt
aiohttp
aiosqlite
discord.py
youtube_dl
yt_dlp
pynacl